New style of scanning and decodingi changes.
Scanning preferences for a plate now just define the region to be scanned.
The grid will now be presented to the user after the plate is scanned. The
user will then be allowed to adjust the grid to get better results from
decoding. This will also alow users to upload an image of a plate instead of
using a flatbed scanner.